:tada: Initial commit
This is nowhere near complete yet; but this was a good pause point at which
to make an initial commit. Mostly I think I have an idea as to how I'm going
to do this in Common Lisp, and at this moment the code is capable of opening
and reading from a 2bit file, loading its index.

No actual access to sequence data is implemented yet; that's up next.