:truck: Rename with-2bit-data to with-2bit-file
The plan with this code is that you don't *have* to pull the data from a
local file; in the future it could be nice to add other sources (via
HTTP/HTTPS, for example -- anything that allows for the reading of a section
of the source file). The with-2bit-data macro was named with this in mind,
but internally it actually assumes that we're working with a file.

Here I rename it to reflect the fact that it's all about files. Later I'll
add macros for any other source that comes along; and perhaps then add
something that works out the best form to use.